requirements.txt
pyaudio
pygame
speechrecognition
python-dotenv
newsapi-python
openai
pyautogui
pvporcupine
sounddevice